yaml-generator/requirements.txt
watchdog
pyyaml
groq